SHALLOW CLONE support for Delta tables with deletion vectors
This PR is part of the feature: Support Delta tables with deletion vectors (more details at delta-io/delta#1485)

This PR adds support for SHALLOW CLONEing a Delta table with deletion vectors. The main change is to convert the relative path of DV file in AddFile to absolute path when cloning the table.

Added tests

Closes delta-io/delta#1733